Enter接口Debug

+++ server/dto/common.go
@@ -1,5 +1,5 @@
 package dto
 
 type ResourceID struct {
-	ID uint32 `json:"id"`
+	ID int64 `json:"id"`
 }
